requirements.txt
numpy==1.19.3
Pillow==8.0.1
opencv-python~=4.4.0.46